+++ cmd/zetae2e/local/evm.go
@@ -95,6 +95,7 @@ func startEVMTests(eg *errgroup.Group, conf config.Config, deployerRunner *runne
 		),
 	)
 
+	// TODO evm: tmp comment out because of setup difference between zetae2e-ante and zetae2e
 	eg.Go(
 		evmTestRoutine(conf, "zeta", conf.AdditionalAccounts.UserZeta, color.FgHiBlue, deployerRunner, verbose,
 			e2etests.TestZetaDepositName,