+++ e2e/e2etests/test_sui_withdraw_and_call_revert_with_call_legacy.go
@@ -56,7 +56,7 @@ func TestSuiWithdrawAndCallRevertWithCallLegacy(r *runner.E2ERunner, args []stri
 			OnRevertGasLimit: big.NewInt(0),
 		},
 	)
-	r.Logger.EVMTransaction(*tx, "withdraw_and_call")
+	r.Logger.EVMTransaction(tx, "withdraw_and_call")
 
 	// ASSERT
 	// wait for the cctx to be reverted