+++ x/fungible/types/evm_test.go
@@ -5,9 +5,9 @@ import (
 	"fmt"
 	"testing"
 
+	evmtypes "github.com/cosmos/evm/x/vm/types"
 	"github.com/ethereum/go-ethereum/core/vm"
 	"github.com/stretchr/testify/require"
-	evmtypes "github.com/zeta-chain/ethermint/x/evm/types"
 
 	"github.com/zeta-chain/node/x/fungible/types"
 )